프로젝트 구조 및 파일 수정
