+++ src/utils/common/changeFromCharToDecimal.ts
@@ -1,0 +1,6 @@
+// 문자(char)를 10진수(Decimal)로 변환하는 함수
+const changeFromCharToDecimal = (char: string) => {
+  return char.charCodeAt(0);
+};
+
+export default changeFromCharToDecimal;
