+++ src/utils/common/encodeFromBufferToBase64.ts
@@ -1,0 +1,6 @@
+// 문자열을 Base64로 인코딩하는 함수
+const encodeFromBufferToBase64 = (data: string): string => {
+  return Buffer.from(data).toString('base64');
+};
+
+export default encodeFromBufferToBase64;
